Rework translation set selection & creation UI
