+++ lib/modules/translation/screens/translations.dart
@@ -1,169 +1,116 @@
+import 'package:andax/models/node.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/models/translation.dart';
-import 'package:andax/modules/translation/screens/translation_editor.dart';
-import 'package:andax/modules/translation/services/translations.dart';
-import 'package:andax/modules/translation/widgets/translation_creator.dart';
-import 'package:andax/shared/services/story_loader.dart';
+import 'package:andax/modules/editor/utils/editor_sheet.dart';
+import 'package:andax/modules/translation/services/assets.dart';
+import 'package:andax/shared/widgets/danger_dialog.dart';
 import 'package:andax/shared/widgets/loading_dialog.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
-class TranslationsScreen extends StatefulWidget {
-  const TranslationsScreen({
+import '../widgets/asset.dart';
+
+typedef AssetOverwrite = MapEntry<String, String>;
+
+class TranslationEditorScreen extends StatefulWidget {
+  const TranslationEditorScreen({
     required this.info,
+    required this.narrative,
+    required this.base,
+    required this.target,
     Key? key,
   }) : super(key: key);
 
   final StoryInfo info;
+  final Story narrative;
+  final Translation base;
+  final Translation target;
 
   @override
-  State<TranslationsScreen> createState() => _TranslationsScreenState();
+  State<TranslationEditorScreen> createState() => TranslationEditorState();
 }
 
-class _TranslationsScreenState extends State<TranslationsScreen> {
-  List<StoryInfo>? translations;
-  late StoryInfo base;
-  late StoryInfo target;
+class TranslationEditorState extends State<TranslationEditorScreen> {
+  StoryInfo get info => widget.info;
+  Story get narrative => widget.narrative;
+  Translation get base => widget.base;
+  Translation get target => widget.target;
+
+  final changes = <String, AssetOverwrite>{};
 
   @override
-  void initState() {
-    super.initState();
-    getAllTranslations(widget.info.storyID).then(
-      (s) => setState(() {
-        translations = s;
-        base = s.first;
-        target = base;
-      }),
-    );
-  }
-
-  void previewTranslation(StoryInfo info) {
-    showDialog<void>(
-      context: context,
-      builder: (context) {
-        return AlertDialog(
-          title: Text(info.title),
-          content: info.description.isEmpty ? null : Text(info.description),
-          actions: [
-            TextButton.icon(
-              onPressed: () {
-                setState(() => base = info);
-                Navigator.pop(context);
-              },
-              icon: const Icon(Icons.source_rounded),
-              label: const Text('As base'),
-            ),
-            TextButton.icon(
-              onPressed: () {
-                setState(() => target = info);
-                Navigator.pop(context);
-              },
-              icon: const Icon(Icons.translate_rounded),
-              label: const Text('As target'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> openEditor() async {
-    late final Story narrative;
-    late final Translation base;
-    late final Translation target;
-    await showLoadingDialog(
-      context,
-      (() async {
-        narrative = await loadNarrative(this.base);
-        base = await loadTranslation(this.base);
-        target = await loadTranslation(this.target);
-      })(),
-    );
-    await Navigator.push<void>(
-      context,
-      MaterialPageRoute(
-        builder: (context) {
-          return TranslationEditorScreen(
-            info: this.target,
-            narrative: narrative,
-            base: base,
-            target: target,
-          );
-        },
-      ),
-    );
+  void setState(void Function() fn) {
+    super.setState(fn);
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        leading: const RoundedBackButton(),
-        title: const Text('Select trasnlations'),
-        actions: [
-          if (translations != null)
-            IconButton(
-              icon: const Icon(Icons.note_add_rounded),
-              onPressed: () async {
-                final data = await showTranslationCreator(context);
-                if (data != null) {
-                  final info = await showLoadingDialog(
-                    context,
-                    createTranslation(
-                      context,
-                      widget.info,
-                      data.key,
-                      data.value,
-                    ),
-                  );
-                  if (info != null) {
-                    target = info;
-                    await openEditor();
-                    Navigator.pop(context);
-                  }
-                }
-              },
+    return Provider.value(
+      value: this,
+      child: Builder(
+        builder: (context) {
+          return WillPopScope(
+            onWillPop: () => showDangerDialog(
+              context,
+              'Leave editor? Unsaved progress will be lost!',
+              confirmText: 'Exit',
+              rejectText: 'Stay',
             ),
-          const SizedBox(width: 4),
-        ],
-      ),
-      floatingActionButton: translations == null
-          ? null
-          : FloatingActionButton.extended(
-              onPressed: () async {
-                await openEditor();
-                Navigator.pop(context);
-              },
-              icon: const Icon(Icons.edit_note_rounded),
-              label: const Text('Edit translation'),
+            child: Scaffold(
+              appBar: AppBar(
+                leading: const RoundedBackButton(),
+                title: const Text('Translations editor'),
+              ),
+              floatingActionButton: changes.isNotEmpty &&
+                      FirebaseAuth.instance.currentUser?.uid ==
+                          info.translationAuthorID
+                  ? FloatingActionButton(
+                      onPressed: () async {
+                        await showLoadingDialog(
+                          context,
+                          applyAssetChanges(info, changes),
+                        );
+                        Navigator.pop(context);
+                      },
+                      child: const Icon(Icons.upload_rounded),
+                    )
+                  : null,
+              body: ListView(
+                children: [
+                  ListTile(
+                    leading: const Icon(Icons.language_rounded),
+                    title: Text(target.language),
+                  ),
+                  buildExplanationTile(context, 'General info'),
+                  // ignore: prefer_const_constructors
+                  Asset('title', icon: Icons.title_rounded),
+                  // ignore: prefer_const_constructors
+                  Asset('description', icon: Icons.description_rounded),
+                  // ignore: prefer_const_constructors
+                  Asset('tags', icon: Icons.tag_rounded),
+                  buildExplanationTile(context, 'Actors'),
+                  for (final aid in narrative.actors.keys)
+                    Asset(aid, icon: Icons.person_rounded),
+                  buildExplanationTile(context, 'Cells'),
+                  for (final cid in narrative.cells.entries
+                      .where((e) => e.value.display != null)
+                      .map((e) => e.key))
+                    Asset(cid, icon: Icons.article_rounded),
+                  buildExplanationTile(context, 'Narrative'),
+                  for (final n in narrative.nodes.values) ...[
+                    Asset(n.id, icon: Icons.chat_bubble_rounded),
+                    if (n.input == NodeInputType.select)
+                      for (final t in n.transitions)
+                        Asset(t.id, icon: Icons.call_split_rounded),
+                    const Divider(indent: 64),
+                  ],
+                ],
+              ),
             ),
-      body: ListView(
-        children: [
-          if (translations == null)
-            const ListTile(
-              leading: CircularProgressIndicator(),
-              title: Text('Look for translations'),
-              subtitle: Text('Please, wait...'),
-            )
-          else ...[
-            for (final s in translations!)
-              ListTile(
-                leading: s == base
-                    ? const Icon(Icons.source_rounded)
-                    : s == target
-                        ? const Icon(Icons.translate_rounded)
-                        : null,
-                title: Text(s.title),
-                subtitle: s.description.isEmpty
-                    ? null
-                    : Text(
-                        s.description,
-                        maxLines: 1,
-                      ),
-                onTap: () => previewTranslation(s),
-              )
-          ]
-        ],
+          );
+        },
       ),
     );
   }
