+++ lib/modules/translation/services/assets.dart
@@ -1,7 +1,7 @@
 import 'package:andax/models/story.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-import '../screens/translation_editor.dart';
+import '../screens/translations.dart';
 
 Future<List<AssetOverwrite>> getPendingAssets(StoryInfo info, String id) async {
   final q = await FirebaseFirestore.instance
