+++ lib/modules/translation/widgets/asset.dart
@@ -1,9 +1,9 @@
-import 'package:andax/modules/translation/screens/translation_editor.dart';
 import 'package:andax/modules/translation/services/assets.dart';
 import 'package:andax/shared/widgets/loading_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import '../screens/translations.dart';
 import 'asset_editor.dart';
 
 class Asset extends StatelessWidget {
